Add the sun to the simulation
Insert the sun as a body, grabbing an initial position and velocity from
NASA, based on today's date. This ensures that we calculate our initial
velocity correctly to have our object above the surface.

The object is still placed above the celestial north pole of the earth,
as we don't yet account for the rotation of the earth.

Signed-off-by: David Brown <davidb@davidb.org>