Update the Earth's axial tilt
Use a slightly more accurate value for the axial tilt of the Earth. The
values are still a bit inconsistent, as the axial tilt is for J2000 but
the position/velocity is for Sep 23, 2025.  But, realistically, it
doesn't really matter much, as this is unlikely to be used to plan
actual missions.

Signed-off-by: David Brown <davidb@davidb.org>