examples: tennis: Simplification
Instead of converting the angular momentum to body frame and back each
step, just leave it in the body frame. This simplifies the code a lot.
The tradeoff is that there might be slight numerical error, as the
momentum is effectively recalculated. However, this is likely negligible
compared to other numerical errors in the simulation.

Signed-off-by: David Brown <davidb@davidb.org>