sim-physics: Implement Improved algorithm
Based on the paper "Improved quaternion-based integration scheme for
rigid body motion" by L.J.H. Seelen, J.T. Padding, and J.A.M. Kuipers.
https://link.springer.com/article/10.1007/s00707-016-1670-x

Somewhat mindbogglingly, this code is a result of feeding ChatGPT 5
Thinking the previous version of this file, the pdf of the paper, and a
bit of prompting about what I want.

Signed-off-by: David Brown <davidb@davidb.org>